requirements.txt
# Runtime dependencies
Django==3.1.13
mysqlclient==2.0.3
Pillow==8.3.2
sorl-thumbnail==12.6.3
# Dev dependencies
black==21.8b0
